backend/pyproject.toml
[tool.poetry]
name = "wisdom-backend"
version = "0.1.0"
description = "Local FastAPI backend for Wisdom app (no Supabase dependencies)"
authors = ["Wisdom Team <dev@wisdom.local>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }
psycopg = { version = "^3.2.1", extras = ["pool"] }
python-dotenv = "^1.0.1"
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
pydantic-settings = "^2.4.0"
stripe = "^5.0.0"
bcrypt = "<5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
httpx = "^0.27.0"
ruff = "^0.6.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
